2022.1/universal/fix/switch-spotlight to 2022.1/staging
backport https://fogbugz.unity3d.com/f/cases/1422306/
https://unity3d.atlassian.net/browse/TRND-2072

This fixes incorrect distance falloff for spot and point lights on Nintendo Switch.

+++ Packages/com.unity.render-pipelines.universal/Shaders/Utils/Deferred.hlsl
@@ -32,7 +32,8 @@ Light UnityLightFromPunctualLightDataAndWorldSpacePosition(PunctualLightData pun
 
     half3 lightDirection = half3(lightVector * rsqrt(distanceSqr));
 
-    half attenuation = DistanceAttenuation(distanceSqr, punctualLightData.attenuation.xy) * AngleAttenuation(punctualLightData.spotDirection.xyz, lightDirection, punctualLightData.attenuation.zw);
+    // full-float precision required on some platforms
+    float attenuation = DistanceAttenuation(distanceSqr, punctualLightData.attenuation.xy) * AngleAttenuation(punctualLightData.spotDirection.xyz, lightDirection, punctualLightData.attenuation.zw);
 
     light.direction = lightDirection;
     light.color = punctualLightData.color.rgb;
